extract single owner arena implementation into its own class
